requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Session==0.8.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
msgspec==0.18.6
multidict==6.1.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.1
numpy==2.0.2
openai==0.28.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
propcache==0.2.1
prov==2.0.1
puremagic==1.28
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.3
PyMuPDF==1.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0